requirements/dev.txt
black==23.9.1
crispy-bootstrap5==2023.10
django==4.2.5
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
flake8==6.1.0
pillow==10.0.1
python-decouple==3.8
sorl-thumbnail==12.10.0